Reuse tutorial content for the getting started guide
By using the include directive, the tutorial instructions to create a
bootable USB drive and running the Clear Linux installer are reused on
the getting started guide. Moving forward, the instructions can be
updated in the tutorial and they will automatically update on the
getting started. By adding labels to the the tutorial content, I was
able to make the include directive independent of the line numbers.
Thus, even if content is added or removed, the inclusion won't break.

Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>